+++ src/opofea.h
@@ -34,7 +34,9 @@
 #include "common.h"
 
 void run_fea1p1(const string folder_path,
-    shared_ptr<IOHprofiler_suite<int> > suite, const int eval_budget,
-    const int independent_runs, const unsigned rand_seed);
+    shared_ptr<IOHprofiler_suite<int>> suite,
+    const unsigned long long eval_budget,
+    const unsigned long long independent_runs,
+    const unsigned long long rand_seed);
 
 #endif
